Used Cassandra system environment.

+++ api/src/main/java/net/hrobotics/bt/data/Storage.java
@@ -24,9 +24,19 @@ public class Storage {
     public void init() {
         System.out.println("Storage.init - begin");
 
-        Cluster cluster = Cluster.builder().addContactPoint(LOCALHOSTIP).build();
+        String s = System.getenv(BT_CASSANDRA_PORT_9042_TCP_ADDR);
 
-        session = cluster.connect(KEYSPACE);
+        if (s != null) {
+            System.out.println("Storage.init - BT_CASSANDRA_PORT_9042_TCP_ADDR = " + s);
+
+            Cluster cluster = Cluster.builder().addContactPoint(s).build();
+
+            session = cluster.connect(KEYSPACE);
+        } else {
+            System.out.println("Environment variable BT_CASSANDRA_PORT_9042_TCP_ADDR not determined");
+
+            session = null;
+        }
 
         System.out.println("Storage.init - end");
     }
@@ -59,18 +69,24 @@ public class Storage {
     }
 
     public void deleteUserEmail(String email) {
+        if (session == null) { return; }
+
         Statement delete = QueryBuilder.delete().from(KEY_USEREMAIL).where(QueryBuilder.eq(KEY_EMAIL, email));
 
         session.execute(delete);
     }
 
     public void insertUserEmail(String email, UUID uuid) {
+        if (session == null) { return; }
+
         Statement insert = QueryBuilder.insertInto(KEY_USEREMAIL).value(KEY_EMAIL, email).value(KEY_USERID, uuid);
 
         session.execute(insert);
     }
 
     public UUID selectUserEmail(String email) {
+        if (session == null) { return null; }
+
         Statement select = QueryBuilder.select().all().from(KEY_USEREMAIL).where(QueryBuilder.eq(KEY_EMAIL, email));
 
         ResultSet rs = session.execute(select);
@@ -83,6 +99,8 @@ public class Storage {
     }
 
     public void updateUserEmail(String email, UUID uuid) {
+        if (session == null) { return; }
+
         Statement update = QueryBuilder.update(KEY_USEREMAIL).with(QueryBuilder.set(KEY_USERID, uuid))
                                        .where(QueryBuilder.eq(KEY_EMAIL, email));
 
@@ -90,11 +108,11 @@ public class Storage {
     }
 
 
-    //private static final String LOCALHOSTIP   = "127.0.0.1";
-    private static final String LOCALHOSTIP   = "172.17.0.2";
     private static final String KEYSPACE      = "btkeyspace";
 
     private static final String KEY_EMAIL     = "email";
     private static final String KEY_USEREMAIL = "useremail";
     private static final String KEY_USERID    = "userid";
+
+    private static final String BT_CASSANDRA_PORT_9042_TCP_ADDR = "BT_CASSANDRA_PORT_9042_TCP_ADDR";
 }
